+++ MyMoney.Web/ClientApp/src/app/app.module.ts
@@ -25,11 +25,13 @@ import {
    ResetPasswordComponent,
    ChangePasswordComponent,
    EditRecurringTransactionsComponent,
+   AddBasicTransactionsComponent,
 } from './pages';
 import {
    ChartComponent,
    FooterComponent,
-   ImportFileComponent
+   ImportFileComponent,
+   ToggleComponent
 } from './shared/components';
 import { StoreModule } from '@ngrx/store';
 import { appReducer } from './shared/state/app-state';
@@ -61,11 +63,13 @@ import { routes } from './app.routing';
       ResetPasswordComponent,
       ChangePasswordComponent,
       EditRecurringTransactionsComponent,
+      AddBasicTransactionsComponent,
 
       // Shared
       ImportFileComponent,
       ChartComponent,
       FooterComponent,
+      ToggleComponent,
    ],
    imports: [
       BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
