+++ MyMoney.Web/ClientApp/src/app/pages/index.ts
@@ -17,3 +17,4 @@ export * from './auth/forgot-password/forgot-password.component';
 export * from './auth/reset-password/reset-password.component';
 export * from './user/change-password/change-password.component';
 export * from './transactions/edit-recurring/edit-recurring.transactions.component';
+export * from './transactions/add/add-basic/add-basic.transactions.component';
