+++ MyMoney.Web/ClientApp/src/app/pages/transactions/add/add-basic/add-basic.transactions.component.html
@@ -1,0 +1,89 @@
+<div class="row">
+   <div class="col-6">
+      <form [formGroup]="addTransactionForm" (ngSubmit)="onSubmit()">
+         <div class="form-group">
+            <label for="date">Date</label>
+            <input type="date" formControlName="date" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.date.errors }"
+               (change)="onDateChange()" />
+            <div *ngIf="submitted && f.date.errors" class="invalid-feedback">
+               <div *ngIf="f.date.errors.required">Date is required</div>
+            </div>
+         </div>
+         <div class="form-group">
+            <label for="description">Description</label>
+            <input type="text" formControlName="description" class="form-control"
+               [ngClass]="{ 'is-invalid': submitted && f.description.errors }" />
+            <div *ngIf="submitted && f.description.errors" class="invalid-feedback">
+               <div *ngIf="f.description.errors.required">Description is required</div>
+            </div>
+         </div>
+         <div class="form-group">
+            <label for="amount">Amount</label>
+            <input type="number" formControlName="amount" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.amount.errors }" />
+            <div *ngIf="submitted && f.amount.errors" class="invalid-feedback">
+               <div *ngIf="f.amount.errors.required">Amount is required</div>
+               <div *ngIf="f.amount.errors.min">Amount cannot be negative</div>
+            </div>
+         </div>
+         <div class="form-group">
+            <label for="notes">Notes</label>
+            <textarea formControlName="notes" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.notes.errors }"
+               rows="3"></textarea>
+            <div *ngIf="submitted && f.notes.errors" class="invalid-feedback"></div>
+         </div>
+         <div class="form-group">
+            <button [disabled]="loading" class="btn btn-success">
+               <span *ngIf="loading" class="spinner-border spinner-border-sm mr-1"></span>
+               Add
+            </button>
+            <a routerLink="/transactions" class="btn btn-link">Cancel</a>
+         </div>
+      </form>
+   </div>
+   <div class="col-6">
+      <h4>Budgets</h4>
+      <table class="table">
+         <thead>
+            <tr>
+               <th scope="col"></th>
+               <th scope="col">Name</th>
+               <th scope="col">Amount</th>
+               <th scope="col">Remaining</th>
+            </tr>
+         </thead>
+         <tbody>
+            <tr *ngFor="let budget of budgets; let i = index">
+               <td>
+                  <input type="checkbox" [value]="budget.id" [checked]="selectedBudgets.has(budget.id)"
+                     (change)="onBudgetCheckboxChange($event, budget.id)" />
+               </td>
+               <td>{{budget.name}}</td>
+               <td>{{budget.amount}}</td>
+               <td [ngClass]="{ 'text-danger': budget.showWarning }">{{budget.remaining}}</td>
+            </tr>
+         </tbody>
+      </table>
+      <h4>Incomes</h4>
+      <table class="table">
+         <thead>
+            <tr>
+               <th scope="col"></th>
+               <th scope="col">Date</th>
+               <th scope="col">Name</th>
+               <th scope="col">Remaining</th>
+            </tr>
+         </thead>
+         <tbody>
+            <tr *ngFor="let income of incomes; let i = index">
+               <td>
+                  <input type="checkbox" [value]="income.id" [checked]="selectedIncomes.has(income.id)"
+                     (change)="onIncomeCheckboxChange($event, income.id)" />
+               </td>
+               <td>{{income.date}}</td>
+               <td>{{income.name}}</td>
+               <td [ngClass]="{ 'text-danger': income.showWarning }">{{income.remaining}}</td>
+            </tr>
+         </tbody>
+      </table>
+   </div>
+</div>
