+++ MyMoney.Web/ClientApp/src/app/pages/transactions/add/add.transactions.component.scss
@@ -1,0 +1,6 @@
+.header {
+   display: flex;
+   flex-direction: row;
+   width: 100%;
+   justify-content: space-between;
+}
