+++ MyMoney.Web/ClientApp/src/app/shared/components/index.ts
@@ -1,3 +1,4 @@
 export * from './chart';
 export * from './footer/footer.component';
 export * from './import';
+export * from './toggle/toggle.component';
