+++ MyMoney.Web/ClientApp/src/app/shared/components/toggle/toggle.component.html
@@ -1,0 +1,9 @@
+<div class="toggle-container">
+   <label class="switch">
+      <input type="checkbox" [checked]="value" (change)="onChecked($event)">
+      <span class="slider round"></span>
+   </label>
+   <div class="label">
+      {{ text }}
+   </div>
+</div>
